refactor: Revised project structure for improved organization.
