+++ api/queryFunctions.js
