+++ app/hotels/[hotelId]/api/mockData.js
