+++ app/hotels/components/HotelList.jsx
@@ -4,7 +4,7 @@ import { observer } from "@legendapp/state/react";
 import { HotelCard } from "./HotelCard";
 import { appliedFilters$, sortingOptions$ } from "../store";
 import { filterHotels, sortHotels } from "../utils";
-import { useGetFilteredHotelsFromURL } from "../hooks/useGetFilteredHotelsFromURL";
+import { useGetFilteredHotelsFromURL } from "../api/useGetFilteredHotelsFromURL";
 
 // TODO: Extract duration dynamically from URL
 const STAY_DURATION_DAYS = 5;
